refactor: move logic from database to user service

+++ src/post/user/user.service.ts
@@ -1,12 +1,31 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import { DataService } from 'src/common/services';
 import { CreateUserDto, UpdatePasswordDto } from 'src/common/interfaces';
+import {
+  isCreateUserData,
+  isUpdateUserData,
+  statusResponse,
+} from 'src/common/utils';
+import { randomUUID } from 'crypto';
+import { isUUID } from 'class-validator';
 
 @Injectable()
 export class UserService {
   constructor(private dataBase: DataService) {}
+
   create(createUserDto: CreateUserDto) {
-    return this.dataBase.createUser(createUserDto);
+    if (!isCreateUserData(createUserDto)) {
+      return statusResponse(HttpStatus.BAD_REQUEST);
+    }
+    const dto = {
+      id: randomUUID({ disableEntropyCache: true }),
+      version: 1,
+      createdAt: Date.now(),
+      updatedAt: Date.now(),
+      login: createUserDto.login,
+      password: createUserDto.password,
+    };
+    return this.dataBase.createUser(dto);
   }
 
   findAll() {
@@ -14,14 +33,24 @@ export class UserService {
   }
 
   findOne(id: string) {
-    return this.dataBase.getUser(id);
+    if (!isUUID(id, '4')) return statusResponse(HttpStatus.BAD_REQUEST);
+    const res = this.dataBase.getUser(id);
+    return res ? res : statusResponse(HttpStatus.NOT_FOUND);
   }
 
   update(id: string, updatePasswordDto: UpdatePasswordDto) {
-    return this.dataBase.updateUser(id, updatePasswordDto);
+    if (!isUUID(id, '4')) return statusResponse(HttpStatus.BAD_REQUEST);
+    if (!isUpdateUserData(updatePasswordDto)) {
+      return statusResponse(HttpStatus.BAD_REQUEST);
+    }
+    const updateUser = this.dataBase.updateUser(id, updatePasswordDto);
+    if (updateUser === null) return statusResponse(HttpStatus.FORBIDDEN);
+    return updateUser ? updateUser : statusResponse(HttpStatus.NOT_FOUND);
   }
 
   remove(id: string) {
-    return this.dataBase.deleteUser(id);
+    if (!isUUID(id, '4')) return statusResponse(HttpStatus.BAD_REQUEST);
+    const user = this.dataBase.deleteUser(id);
+    return user ? user : statusResponse(HttpStatus.NOT_FOUND);
   }
 }
